chore(admin-service): move to packages
